Changed the artifact ID to io.jenkins.plugins
