+++ src/main/java/io/jenkins/plugins/Client/IClient.java
@@ -1,0 +1,7 @@
+package io.jenkins.plugins.Client;
+
+import java.util.HashMap;
+
+public interface IClient {
+    boolean transmitData(HashMap<String, Object> payload, boolean isEvent);
+}
