+++ src/main/java/io/jenkins/plugins/Metrics/interfaces/IPublishMetrics.java
@@ -1,0 +1,7 @@
+package io.jenkins.plugins.Metrics.interfaces;
+
+import java.util.HashMap;
+
+public interface IPublishMetrics {
+    HashMap<String, Object> collectMetrics(Object details);
+}
