--- src/main/java/org/maplelabs/jenkins/plugins/Client/IClient.java
@@ -1,7 +1,0 @@
-package org.maplelabs.jenkins.plugins.Client;
-
-import java.util.HashMap;
-
-public interface IClient {
-    boolean transmitData(HashMap<String, Object> payload, boolean isEvent);
-}
