--- src/main/java/org/maplelabs/jenkins/plugins/Metrics/interfaces/IPublishMetrics.java
@@ -1,7 +1,0 @@
-package org.maplelabs.jenkins.plugins.Metrics.interfaces;
-
-import java.util.HashMap;
-
-public interface IPublishMetrics {
-    HashMap<String, Object> collectMetrics(Object details);
-}
